+++ examples/create_order.py
@@ -4,8 +4,8 @@ Sample for creating an order on the derive client.
 
 from rich import print
 
-from derive.derive import DeriveClient
-from derive.enums import Environment, OrderSide
+from derive_client.derive import DeriveClient
+from derive_client.enums import Environment, OrderSide
 from tests.conftest import TEST_PRIVATE_KEY
