+++ tests/test_user.py
@@ -7,7 +7,7 @@ from time import sleep
 
 import pytest
 
-from derive.enums import OrderSide, OrderType
+from derive_client.enums import OrderSide, OrderType
 
 
 @pytest.mark.skip(reason="This test is not meant to be run in CI")
